refactor: Car 가 아닌 CarName 에서 자동차 이름 검증
