+++ src/main/java/racinggame/domain/Cars.java
@@ -4,7 +4,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Cars {
-	static final String ERROR_MESSAGE = "[ERROR] 자동차의 이름은 5자 이하입니다. 다시 입력해주세요.";
 	private final List<Car> cars = new ArrayList<>();
 
 	private Cars(List<String> carNames) {
