+++ src/main/java/racinggame/view/CarInput.java
@@ -5,7 +5,7 @@ import java.util.List;
 import java.util.NoSuchElementException;
 
 import nextstep.utils.Console;
-import racinggame.domain.Car;
+import racinggame.domain.CarName;
 
 public class CarInput {
 	public static final String GUIDE_MESSAGE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
@@ -24,7 +24,7 @@ public class CarInput {
 		System.out.println(GUIDE_MESSAGE);
 		List<String> inputs = Arrays.asList(Console.readLine().split(DELIMITER));
 		for (String input : inputs)
-			if (Car.isOverThanFiveLetters(input))
+			if (CarName.isOverThanFiveLetters(input))
 				return getInput();
 		return inputs;
 	}
